requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scipy>=1.10.0
spyder>=5.4.0
plotly>=5.13.0
PyQt